backend/requirements.txt

fastapi
uvicorn[standard]
asyncpg
PyJWT
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-multipart
databases[asyncpg]
sqlalchemy
sqlalchemy[asyncio]
pydantic-settings
alembic
pydantic
redis>=4.5.0
httpx
python-dotenv
weaviate-client>=4.6.0,<5.0.0
openai
redis[asyncio]
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
PyPDF2>=3.0.0
psycopg2-binary
numpy
requests
google-auth

